+++ benchmarks/BenchmarkPost.php
@@ -7,11 +7,12 @@ namespace JacobJoergensen\LaravelPaper\Benchmarks;
 use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
+use JacobJoergensen\LaravelPaper\Contracts\PaperModel;
 use JacobJoergensen\LaravelPaper\Paper;
 
 #[Driver('markdown')]
 #[ContentPath('benchmarks/.fixtures/posts')]
-final class BenchmarkPost extends Model
+final class BenchmarkPost extends Model implements PaperModel
 {
     use Paper;