+++ tests/Fixtures/Page.php
@@ -7,11 +7,12 @@ namespace JacobJoergensen\LaravelPaper\Tests\Fixtures;
 use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
+use JacobJoergensen\LaravelPaper\Contracts\PaperModel;
 use JacobJoergensen\LaravelPaper\Paper;
 
 #[Driver('json')]
 #[ContentPath('tests/content/pages')]
-final class Page extends Model
+final class Page extends Model implements PaperModel
 {
     use Paper;
 }